requirements.txt
aiogram>=3.7.0,<4.0.0
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
asyncpg>=0.29.0,<1.0.0
pydantic-settings>=2.3.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
PyJWT>=2.9.0,<3.0.0
bcrypt>=4.1.3,<5.0.0
openpyxl>=3.1.0,<4.0.0
